The Bad Moon Loonshirne rules are fixed: - The new rules are not part of the subfactions, so if your General have any of the Keywords the rules must be applied. - The SQUIG rules are optional, you can use the Moonclan rules if you want. - The TROGGOTH and SPIDERFANG rules are mandatory to be applied when you select a GENERAL with that Keyword.
The Loonboss with a Giant Cave Squig was triggering all the rules with the SQUIG Keyword without having that Keyword. This problem was fixed and he only triggers the Battleline units.

The Troggoth Hag have the correct artefacts and command traits when she is the general in any of the subfactions.

You can include any ARACHNAROK in the Grimscuttle Spider Cluster Battalion

Added Fellwater and Rockgut Keywords.